Endgame
I added the stuff that allows you to win into the cabin hallway, but
you feel free to tweak it however you'd like so it functions with the
actual winning of the game stuff.  Anyway, I believe I've done
everything that was expected of me, and I might not be around to make
any more changes.