Adding Dynamic Commands
I'm just adding Dynamic Commands as a better way to handle the item
commands.  Whenever the user runs a command it consults the list of
possible dynamic commands and builds a list of temporary commands that
are possible based on the verb given and any items present in the room,
and then it checks to see if one of these commands happens to be the
one you've given.