Dynamic Commands
I made it so that Dynamic Commands are separated by universale and
room, like the rest of the commands.  Now the order of importance for
commands is: Dynamic Room, Static Room, Wildcard Room, Dynamic
Fallback, Static Fallback, Wildcard Fallback, and of course as usual if
these all fail it gives the standard response.